move blueprints under sub-package
