+++ .gitignore
@@ -16,6 +16,7 @@ develop-eggs
 lib
 lib64
 dist/
+build/
 
 # Installer logs
 pip-log.txt
