+++ scout/blueprints/core/constants.py
