+++ scout/blueprints/core/templates/_macros.html
