+++ scout/blueprints/core/templates/case.html
