+++ scout/blueprints/core/templates/cases.html
