+++ scout/blueprints/core/templates/gene-panel.html
