+++ scout/blueprints/core/templates/variant.html
