+++ scout/blueprints/core/templates/variants.html
