+++ scout/blueprints/core/views.py
@@ -12,10 +12,10 @@ import query_phenomizer
 from scout.models import Case, Variant
 from scout.models.variant import SEVERE_SO_TERMS
 from scout.extensions import mail, store, loqusdb
-from scout.helpers import templated
+from scout.utils import templated, validate_user
 
 from .forms import init_filters_form, process_filters_form, GeneListUpload
-from .utils import validate_user, genecov_links
+from .utils import genecov_links
 from .constants import PHENOTYPE_GROUPS
 
 core = Blueprint('core', __name__, template_folder='templates')
