+++ scout/blueprints/frontend/__init__.py
