+++ scout/blueprints/frontend/static/scout-logo.png
