+++ scout/blueprints/frontend/templates/index.html
