+++ scout/blueprints/frontend/views.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, unicode_literals
 from flask import Blueprint, render_template, redirect, url_for
 from flask_login import current_user
 
-from .. import __version__
+from scout import __version__
 
 frontend = Blueprint('frontend', __name__, template_folder='templates',
                      static_folder='static',
