+++ scout/blueprints/pileup/partial.py
