+++ scout/utils/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
+from .helpers import templated, validate_user
 from .md5 import generate_md5_key
